fix(pg): read sample chugsplash file instead of copying from source
