+++ .changeset/empty-clocks-sleep.md
@@ -1,0 +1,5 @@
+---
+'@chugsplash/plugins': patch
+---
+
+Fix chugsplash-init bug where task attempted to copy from non-existent file
