--- packages/plugins/src/sample-project/hello-chugsplash.js
@@ -1,18 +1,0 @@
-require('@chugsplash/core')
-
-module.exports = {
-  options: {
-    projectName: 'Hello ChugSplash',
-  },
-  contracts: {
-    MyFirstContract: {
-      contract: 'HelloChugSplash',
-      variables: {
-        number: 1,
-        stored: true,
-        storageName: 'First',
-        otherStorage: '0x1111111111111111111111111111111111111111',
-      },
-    },
-  },
-}
