--- packages/plugins/src/sample-project/hello-chugsplash.ts
@@ -1,20 +1,0 @@
-import { UserChugSplashConfig } from '@chugsplash/core'
-
-const config: UserChugSplashConfig = {
-  options: {
-    projectName: 'Hello ChugSplash',
-  },
-  contracts: {
-    MyFirstContract: {
-      contract: 'HelloChugSplash',
-      variables: {
-        number: 1,
-        stored: true,
-        storageName: 'First',
-        otherStorage: '0x1111111111111111111111111111111111111111',
-      },
-    },
-  },
-}
-
-export default config
