+++ packages/plugins/src/sample-project/index.ts
@@ -1,2 +1,3 @@
 export * from './sample-contract'
 export * from './sample-tests'
+export * from './sample-chugsplash-files'
