+++ packages/plugins/src/sample-project/sample-chugsplash-files.ts
@@ -1,0 +1,41 @@
+export const sampleChugSplashFileTypeScript = `import { UserChugSplashConfig } from '@chugsplash/core'
+
+const config: UserChugSplashConfig = {
+  options: {
+    projectName: 'Hello ChugSplash',
+  },
+  contracts: {
+    MyFirstContract: {
+      contract: 'HelloChugSplash',
+      variables: {
+        number: 1,
+        stored: true,
+        storageName: 'First',
+        otherStorage: '0x1111111111111111111111111111111111111111',
+      },
+    },
+  },
+}
+
+export default config
+`
+
+export const sampleChugSplashFileJavaScript = `require('@chugsplash/core')
+
+module.exports = {
+  options: {
+    projectName: 'Hello ChugSplash',
+  },
+  contracts: {
+    MyFirstContract: {
+      contract: 'HelloChugSplash',
+      variables: {
+        number: 1,
+        stored: true,
+        storageName: 'First',
+        otherStorage: '0x1111111111111111111111111111111111111111',
+      },
+    },
+  },
+}
+`
